requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
djangorestframework==3.11.0
drf-httpsig==1.2.1.dev0+ga84326d.d20200706
future==0.18.2
httpsig==1.3.0
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
pyblake2==1.1.2
pycryptodome==3.9.8
pycryptodomex==3.9.8
python-bitcoinlib==0.11.1.dev0
python-zcashlib===0.7.1-SNAPSHOT
pytz==2020.1
pyzmq==19.0.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
style==1.1.0
update==0.0.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
